+++ app/src/main/java/com/company/rentafield/data/repositories/booking/RemoteUserBookingImpl.kt
@@ -1,6 +1,6 @@
 package com.company.rentafield.data.repositories.booking
 
-import com.company.rentafield.data.datasource.RetrofitService
+import com.company.rentafield.data.services.RetrofitService
 import com.company.rentafield.utils.handleApi
 
 class RemoteUserBookingImpl(
