+++ app/src/main/java/com/company/rentafield/data/repositories/remoteuser/user/RemoteUserRepositoryImpl.kt
@@ -1,7 +1,7 @@
 package com.company.rentafield.data.repositories.remoteuser.user
 
 
-import com.company.rentafield.data.datasource.RetrofitService
+import com.company.rentafield.data.services.RetrofitService
 import com.company.rentafield.domain.DataState
 import com.company.rentafield.utils.handleApi
 import kotlinx.coroutines.flow.Flow
