+++ app/src/main/java/com/company/rentafield/data/services/RetrofitService.kt
@@ -1,4 +1,4 @@
-package com.company.rentafield.data.datasource
+package com.company.rentafield.data.services
 
 import okhttp3.MultipartBody
 import retrofit2.Response
